requirements.txt
loguru==0.7.0
python-telegram-bot==20.4
psycopg2==2.9.6
pydantic~=2.1.1